checkpoint: extract the persistent contract to api/checkpoint
Move the persistent-checkpoint contract — the persisted document types
(Metadata, CheckpointSummary, Summary, Attribution, ...), the option types
(WriteOptions/UpdateOptions/PrecomputedTranscriptBlobs), the reader/writer
interfaces, the Write request union, the sentinel errors, and the
CheckpointVersionBranchV1 const — into a new github.com/entireio/cli/api/checkpoint
package, born with the persistent vocabulary.

The contract depends only on leaf packages (agent/types, checkpoint/id, redact,
go-git plumbing), so a storage backend can implement it without the CLI's
agent/TUI/git machinery. The git implementation (GitStore, Open, the Stores
facade, AuthorReader, and the ephemeral shadow-branch surface) stays in
cmd/entire/cli/checkpoint, which imports the contract and re-exports every
moved symbol via aliases (aliases.go) — all existing call sites compile
unchanged.

Notes:
- PrecomputedTranscriptBlobs.isUsable is now exported (IsUsable); its one
  caller is across the new package boundary.
- The Write union is sealed to api/checkpoint, so the impl-package
  unknown-request test was removed (no longer expressible); per-request
  dispatch coverage remains.
- normalizeCheckpointSummary stays in the impl (read-time normalization).

Entire-Checkpoint: a808b47c0633
