+++ api/checkpoint/errors.go
@@ -1,0 +1,15 @@
+package checkpoint
+
+import "errors"
+
+// Errors returned by checkpoint operations.
+var (
+	// ErrCheckpointNotFound is returned when a checkpoint ID doesn't exist.
+	ErrCheckpointNotFound = errors.New("checkpoint not found")
+
+	// ErrNoTranscript is returned when a checkpoint exists but has no transcript.
+	ErrNoTranscript = errors.New("no transcript found for checkpoint")
+)
+
+// CheckpointVersionBranchV1 identifies the branch-backed checkpoint metadata format.
+const CheckpointVersionBranchV1 = "branch-v1"
